Removed database stuff. Cleaned up code. Removed unused classes. Renamed class to CreateSession
