+++ TrainingApp/app/src/main/java/com/traininapp/View/CreateSession.java
@@ -2,86 +2,106 @@ package com.traininapp.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.DialogFragment;
+import androidx.fragment.app.FragmentTransaction;
 
 import android.app.DatePickerDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
-import android.widget.Button;
 import android.widget.DatePicker;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.traininapp.MainActivity;
-import com.traininapp.Model.DatabaseHelper;
-import com.traininapp.Model.Planning.CardioExercise;
-import com.traininapp.Model.Repository;
 import com.traininapp.Model.Planning.Exercise;
-import com.traininapp.Model.Planning.Routine;
-import com.traininapp.Model.Planning.StrengthExercise;
+import com.traininapp.Model.Repository;
 import com.traininapp.R;
 
-import java.text.DateFormat;
-import java.text.DecimalFormat;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.List;
 
 public class CreateSession extends AppCompatActivity implements DatePickerDialog.OnDateSetListener {
 
-   private List<Routine> listOfAddedRoutines = new ArrayList<>();
-
-   private LocalDate selectedDate;
-
-   private Repository repository;
-   private DatabaseHelper myDB = new DatabaseHelper(this);
-
-
-    private Spinner spnPickRoutine;
-    private Button btnPickDate;
-    private Button btnOk;
-    private Button btnOpenCreateRoutine;
-    private Button btnAddRoutine;
-    private Button btnUndo;
-    private TextView txtDisplayRoutines;
+    // Declaring elements
     private EditText txtEnterSessionName;
+    private TextView txtSelectedDate;
+    private ImageView imgSessionIcon;
+    private int image;
+
+    // Lists for created fragments
+    private List<FragStrRow> listStrFrag = new ArrayList<>();
+    private List<FragCarRow> listCarFrag = new ArrayList<>();
+
+    // List for exercises
+    private List<Exercise> exerciseList = new ArrayList<>();
+
+    // Date of session, set to today's date by default
+    private LocalDate selectedDate = LocalDate.now();
+
+    // Repo
+    private Repository repository;
 
 
+    private FragmentTransaction fragmentTransaction;
+    private boolean control;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_create_session);
+        setContentView(R.layout.activity_create_session_2);
 
-
+        // Initializing the repository
         repository = Repository.getInstance();
 
-        btnPickDate = findViewById(R.id.btnPickDateID);
-        btnOk = findViewById(R.id.btnOkID);
-        btnUndo = findViewById(R.id.btnUndoID);
-        btnAddRoutine = findViewById(R.id.btnAddRoutineID);
-        btnOpenCreateRoutine = findViewById(R.id.btnOpenCreateRoutineID);
-        spnPickRoutine = findViewById(R.id.spnPickRoutineID);
-        txtDisplayRoutines = findViewById(R.id.txtDisplayRoutinesID);
+        // Initializing elements
+        FloatingActionButton btnDone = findViewById(R.id.btnDoneID);
+        TextView txtAddStrExercise = findViewById(R.id.txtAddStrExerciseID);
+        TextView txtAddCarExercise = findViewById(R.id.txtAddCarExerciseID);
+        Spinner spnrIcon = findViewById(R.id.spnrIconID);
+        String[] iconsStrArray = getResources().getStringArray(R.array.iconsStringArray);
         txtEnterSessionName = findViewById(R.id.txtEnterSessionNameID);
+        txtSelectedDate = findViewById(R.id.txtSelectedDateID);
+        imgSessionIcon = findViewById(R.id.imgSessionIconID);
 
-        isDateSelectedAlready();
+        // Setting image to Strength by default
+        image = getResources().getIdentifier("workout_1", "drawable", getPackageName());
 
-        Button btnPickDate = findViewById(R.id.btnPickDateID);
+        // Updating text to match selectedDate, today's date by default
+        txtSelectedDate.setText(selectedDate.format(DateTimeFormatter.ofLocalizedDate(FormatStyle.LONG)));
 
+        // Setting adapter for Spinner and add Listener
+        spnrIcon.setAdapter(new ArrayAdapter(this, android.R.layout.simple_spinner_item, iconsStrArray));
+        spnrIcon.setOnItemSelectedListener(new SpinnerItemSelectedListener());
 
-        //create and setup adapter
-        ArrayAdapter<Routine> adapter = new ArrayAdapter<Routine>(this, android.R.layout.simple_spinner_dropdown_item, repository.getUser().getRoutineList());
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-        spnPickRoutine.setAdapter(adapter);
+        // Clicking on Add exercise text adds a strength exercise row
+        txtAddStrExercise.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                createStrFragRow();
+            }
+        });
 
+        // Clicking on Add exercise text adds a cardio exercise row
+        txtAddCarExercise.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                createCarFragRow();
+            }
+        });
 
-        btnPickDate.setOnClickListener(new View.OnClickListener() {
+        // Clicking the Date button, directs the user to a calendar
+        txtSelectedDate.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
                 DialogFragment datePicker = new DatePickerFragment();
@@ -89,46 +109,35 @@ public class CreateSession extends AppCompatActivity implements DatePickerDialog
             }
         });
 
-        btnAddRoutine.setOnClickListener(new View.OnClickListener() {
+        // Clicking the Done button, saving Session and directing the user to the Upcoming session view
+        btnDone.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View view) {
-                getSelectedRoutine();
-
+            public void onClick(View v) {
+                onDoneClick();
             }
         });
 
-        btnOpenCreateRoutine.setOnClickListener(new View.OnClickListener() {
+        // Clicking the Session icon, directs user to icon selection
+        imgSessionIcon.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View view) {
-                openRoutine();
+            public void onClick(View v) {
+                onIconClick();
             }
         });
-
-
-        btnOk.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                //  Session session = new Session(txtEnterSessionName.getText().toString(), listOfAddedRoutines, )
-                clickSaveSession();
-            }
-        });
-
-        btnUndo.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-
-                undo();
-            }
-        });
-
-
     }
 
     /**
-     * A method to update a string to match the selected date
-     * @param datePicker ???
-     * @param year Year
-     * @param month Month
+     * Method used when pressing Session icon, allows user to choose an icon for the Session
+     */
+    private void onIconClick() {
+    }
+
+    /**
+     * Method which updates selectedDate based on user input
+     *
+     * @param datePicker -
+     * @param year       Year
+     * @param month      Month
      * @param dayOfMonth Day of month
      */
     @Override
@@ -137,132 +146,200 @@ public class CreateSession extends AppCompatActivity implements DatePickerDialog
         c.set(Calendar.YEAR, year);
         c.set(Calendar.MONTH, month);
         c.set(Calendar.DAY_OF_MONTH, dayOfMonth);
-        String currentDateString = DateFormat.getDateInstance().format(c.getTime());
-
 
         // Updating selectedDate to the date selected by user
-        selectedDate = LocalDate.of(year, month, dayOfMonth);
-        setDate(currentDateString,selectedDate);
-    }
-
-    public void setDate(String currentDateString, LocalDate date){
-
-        // Updating selectedDate to the date selected by user
-        selectedDate = date;
-        // Updating the string shown to the date the user selected
-        TextView textView = findViewById(R.id.txtDisplayDateID);
-        textView.setText(currentDateString);
-
-    }
-
-    //Open CreateSession2
-    public void openRoutine(){
-        Intent intent = new Intent(this, CreateSession2.class);
-        startActivity(intent);
+        selectedDate = LocalDate.of(year, month+1, dayOfMonth);
+        txtSelectedDate.setText(selectedDate.format(DateTimeFormatter.ofLocalizedDate(FormatStyle.LONG)));
     }
 
     /**
-     * Method of what happens after clicking the "Done" button. Directs the user to the
-     * Upcoming sessions view
+     * Method called when user press Done button. Saves the information as a new Session and
+     * directs the user to the Upcoming Session view
      */
-    public void clickSaveSession(){
+    public void onDoneClick() {
 
-        // TODO Make this work
-        // Adding the selected Session to the User's Planner's list of Sessions
-        //addSessionToList(selectedRoutine,selectedDate);
+        // Get the Session name from user input
+        String sessionName = txtEnterSessionName.getText().toString();
 
-        // Directing the user to UpcomingSession page again
+        // Remove all previously added exercises
+        exerciseList.clear();
+
+        // Set boolean control true, used for seeing if
+        // everything goes okay when saving the routine
+        control = true;
+
+        // Using the fragments created to save exercises in exerciseList
+        fragStrToStrExList();
+        fragCarToCarExList();
+
+        // Controlling if valid user input
+        //checkName(sessionName);
+        checkNameLength(sessionName);
+
+        // If no fragments returned null and name is unique
+        if (control) {
+
+            // Adding Session to users list
+            repository.getUser().getPlanner().addSession(sessionName, selectedDate, exerciseList, image);
+
+            // Give feedback that the routine has been saved
+            String toastMessage = "Session: " + sessionName + " has been saved!";
+            Toast.makeText(CreateSession.this, toastMessage, Toast.LENGTH_SHORT).show();
+
+            // Clear Session name field
+            txtEnterSessionName.setText("");
+        }
+
+        // Directing the user to Upcoming session view
         Intent intent = new Intent(this, MainActivity.class);
         startActivity(intent);
     }
 
+    /**
+     * Method which adds another row fragment, allowing the user to add exercises
+     */
+    public void createCarFragRow() {
 
-    //Add the routine
-    public void getSelectedRoutine(){
+        // Create the fragment
+        FragCarRow fragment;
 
-        //Get the selected routine
-        Routine addedRoutine = (Routine) spnPickRoutine.getSelectedItem();
+        //Begin the transaction, to start doing something with the fragment
+        fragmentTransaction = getSupportFragmentManager().beginTransaction();
 
-        //Add the routine
-        listOfAddedRoutines.add(addedRoutine);
+        fragment = new FragCarRow();
+        listCarFrag.add(fragment);
 
-        //Get the name of the selected routine
-        String routineName = listOfAddedRoutines.get(listOfAddedRoutines.size()-1).getName();
+        // Add the created fragment to "displayRowsID"
+        fragmentTransaction.add(R.id.displayCarRowsID, fragment);
 
-        //refresh textview when routine has been added
-        refreshTextView();
-
-        //Give feedback to user that the routine has been added
-        String toast = "Routine: " + routineName + " has been added";
-        Toast.makeText(this, toast, Toast.LENGTH_SHORT).show();
-
+        // Commit and finish the FragmentTransaction
+        fragmentTransaction.commit();
     }
 
-    //Remove the most recently added routine
-    public void undo(){
-        //Check if there is anything to remove
-        if(listOfAddedRoutines.size() > 0){
-            listOfAddedRoutines.remove(listOfAddedRoutines.size()-1);
-            //refresh textview when routine has been removed
-            refreshTextView();
-        } else{
-            Toast.makeText(this, "There is nothing to remove! ", Toast.LENGTH_SHORT).show();
+    /**
+     * Method which adds Strength exercises row fragment, allowing the user to add exercises
+     */
+    public void createStrFragRow() {
+
+        // Create the fragment
+        FragStrRow fragment;
+
+        //Begin the transaction, to start doing something with the fragment
+        fragmentTransaction = getSupportFragmentManager().beginTransaction();
+
+        fragment = new FragStrRow();
+        listStrFrag.add(fragment);
+
+        // Add the created fragment to "displayRowsID"
+        fragmentTransaction.add(R.id.displayStrRowsID, fragment);
+
+        // Commit and finish the FragmentTransaction
+        fragmentTransaction.commit();
+    }
+
+    //if the fragment had been removed remove it from the list of added exercises
+    public void removeDeletedExercises() {
+        if (exerciseList.get(exerciseList.size() - 1).getName() == "REMOVE ME") {
+            exerciseList.remove(exerciseList.size() - 1);
         }
     }
 
-    //Refreshes the textview when routine is added or removed
-    public void refreshTextView(){
+    //if failing to save exercise
+    public void failSave() {
+        exerciseList.clear();
+        control = false;
+    }
 
-        //Used to remove trailing zeroes from doubles
-        DecimalFormat removeZeroes = new DecimalFormat("0.#");
+    /**
+     * Method to check if string is at least one character long. If not, setting control to false
+     *
+     * @param name String input from user
+     */
+    public void checkNameLength(String name) {
+        if (name.length() == 0) {
+            String toastMessage = "No name entered!";
+            Toast.makeText(CreateSession.this, toastMessage, Toast.LENGTH_SHORT).show();
+            control = false;
+        }
+    }
 
-        //Make the textview empty before writing all info
-        txtDisplayRoutines.setText("");
+    /**
+     * Method used for creating strength Exercises based on data in strength row fragments
+     */
+    public void fragStrToStrExList() {
 
-        //Go through each routine that has been added
-        for(Routine routine : listOfAddedRoutines){
-            //Show name of selected routine
-            txtDisplayRoutines.append("Routine: " + routine.getName() + "\n");
+        // Go through the list of all created strength fragments
+        for (FragStrRow fr : listStrFrag) {
 
-            //Go through each exercise in that routine
-            for(Exercise exercise : routine.getSavedExerciseList()){
-                //if the exercise is a strength exercise, show the relevant information
-                if(exercise instanceof StrengthExercise){
-                    txtDisplayRoutines.append("Ex: "+ exercise.getName() +
-                            " | Weight: " + removeZeroes.format(((StrengthExercise) exercise).getWeight()) +
-                            " | Sets: " + ((StrengthExercise) exercise).getSets() +
-                            " | Reps: " + ((StrengthExercise) exercise).getReps() +
-                            "\n");
-                }
-                //if the exercise is a cardio exercise, show the relevant information
-                else if(exercise instanceof CardioExercise){
-                    txtDisplayRoutines.append("Ex: "+ exercise.getName() +
-                            " | Dist: " + removeZeroes.format(((CardioExercise) exercise).getDistance()) +
-                            " | Time: " + removeZeroes.format(((CardioExercise) exercise).getRunningTime()) +
-                            "\n");
-                //if the exercise neither, just show the name
-                } else {
-                    txtDisplayRoutines.append("Ex: "+ exercise.getName() + "\n");
-                }
+            // If saveInfo (in FragStrRow) did not return null, add to list of exercises
+            if (fr.saveInfo() != null) {
+                exerciseList.add(fr.saveInfo());
+
+                // Remove exercises of deleted fragments
+                removeDeletedExercises();
+
+                // If it had returned null
+            } else {
+
+                //run failsave and break from loop
+                failSave();
+
+                break;
             }
-
-            //Add extra line between each routine
-            txtDisplayRoutines.append("\n");
         }
     }
 
-    public void isDateSelectedAlready() {
-        Intent intent = getIntent();
-        if (intent.getExtras() != null) {
-            if (intent.getStringExtra("FROMCALENDAR").matches("YES")) {
-                String pastDate = intent.getStringExtra("DATE");
-                DateTimeFormatter formatter = DateTimeFormatter.ofPattern("d - MM - yyyy");
-                LocalDate localDate = LocalDate.parse(pastDate, formatter);
-                setDate(pastDate, localDate);
+    /**
+     * Method used for creating cardio Exercises based on data in cardio row fragments
+     */
+    public void fragCarToCarExList() {
+
+        // Go through the list of all created cardio fragments
+        for (FragCarRow fr : listCarFrag) {
+
+            // If saveInfo (in FragCarRow) did not return null, add to list of exercises
+            // and if control is still true (no strengthexercises returned null)
+            if (fr.saveInfo() != null && control) {
+                exerciseList.add(fr.saveInfo());
+
+                // Remove exercises of deleted fragments
+                removeDeletedExercises();
+
+                // If it had returned null
+            } else {
+
+                // Run failsave and break from loop
+                failSave();
+
+                break;
             }
         }
     }
+
+    /**
+     * Private class used by spinner for setting the Session image
+     */
+    private class SpinnerItemSelectedListener implements AdapterView.OnItemSelectedListener {
+        @Override
+        public void onItemSelected(AdapterView<?> parent, View view,
+                                   int position, long id) {
+
+            // TODO Make this to switch-case. Probably want to generalize setting the image and
+            // TODO getting the image information
+            if (position == 0){
+                imgSessionIcon.setImageResource(R.drawable.workout_1);
+                image = getResources().getIdentifier("workout_1", "drawable", getPackageName());
+            } else {
+                imgSessionIcon.setImageResource(R.drawable.workout_5);
+                image = getResources().getIdentifier("workout_5", "drawable", getPackageName());
+            }
+        }
+
+        @Override
+        public void onNothingSelected(AdapterView<?> parent) {
+            // Empty necessary method
+        }
+    }
+
 
 }
-
-
