+++ TrainingApp/app/src/main/res/layout/activity_create_session_2.xml
@@ -13,7 +13,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="fill_parent"
-            tools:context=".View.CreateSession2">
+            tools:context=".View.CreateSession">
 
 
             <ImageView
